+++ virtcontainers/network.go
@@ -19,12 +19,16 @@ import (
 	"time"
 
 	"github.com/containernetworking/plugins/pkg/ns"
-	"github.com/kata-containers/runtime/virtcontainers/pkg/uuid"
 	"github.com/safchain/ethtool"
 	"github.com/sirupsen/logrus"
 	"github.com/vishvananda/netlink"
 	"github.com/vishvananda/netns"
 	"golang.org/x/sys/unix"
+
+	"github.com/kata-containers/runtime/virtcontainers/device/api"
+	"github.com/kata-containers/runtime/virtcontainers/device/drivers"
+	"github.com/kata-containers/runtime/virtcontainers/pkg/uuid"
+	"github.com/kata-containers/runtime/virtcontainers/utils"
 )
 
 // NetInterworkingModel defines the network model connecting
@@ -261,14 +265,14 @@ func (endpoint *VhostUserEndpoint) SetProperties(properties NetworkInfo) {
 func (endpoint *VhostUserEndpoint) Attach(h hypervisor) error {
 	networkLogger().Info("Attaching vhostuser based endpoint")
 
-	// generate a unique ID to be used for hypervisor commandline fields
-	randBytes, err := generateRandomBytes(8)
+	// Generate a unique ID to be used for hypervisor commandline fields
+	randBytes, err := utils.GenerateRandomBytes(8)
 	if err != nil {
 		return err
 	}
 	id := hex.EncodeToString(randBytes)
 
-	d := VhostUserNetDevice{
+	d := drivers.VhostUserNetDevice{
 		MacAddress: endpoint.HardAddr,
 	}
 	d.SocketPath = endpoint.SocketPath
@@ -331,7 +335,7 @@ func (endpoint *PhysicalEndpoint) Attach(h hypervisor) error {
 		return err
 	}
 
-	d := VFIODevice{
+	d := drivers.VFIODevice{
 		BDF: endpoint.BDF,
 	}
 
@@ -736,7 +740,7 @@ func createFds(device string, numFds int) ([]*os.File, error) {
 	for i := 0; i < numFds; i++ {
 		f, err := os.OpenFile(device, os.O_RDWR, defaultFilePerms)
 		if err != nil {
-			cleanupFds(fds, i)
+			utils.CleanupFds(fds, i)
 			return nil, err
 		}
 		fds[i] = f
@@ -1346,12 +1350,123 @@ func createPhysicalEndpoint(netInfo NetworkInfo) (*PhysicalEndpoint, error) {
 	return physicalEndpoint, nil
 }
 
+// bind/unbind paths to aid in SRIOV VF bring-up/restore
+var pciDriverUnbindPath = "/sys/bus/pci/devices/%s/driver/unbind"
+var pciDriverBindPath = "/sys/bus/pci/drivers/%s/bind"
+var vfioNewIDPath = "/sys/bus/pci/drivers/vfio-pci/new_id"
+var vfioRemoveIDPath = "/sys/bus/pci/drivers/vfio-pci/remove_id"
+
+// bindDevicetoVFIO binds the device to vfio driver after unbinding from host.
+// Will be called by a network interface or a generic pcie device.
+func bindDevicetoVFIO(bdf, hostDriver, vendorDeviceID string) error {
+
+	// Unbind from the host driver
+	unbindDriverPath := fmt.Sprintf(pciDriverUnbindPath, bdf)
+	api.DeviceLogger().WithFields(logrus.Fields{
+		"device-bdf":  bdf,
+		"driver-path": unbindDriverPath,
+	}).Info("Unbinding device from driver")
+
+	if err := utils.WriteToFile(unbindDriverPath, []byte(bdf)); err != nil {
+		return err
+	}
+
+	// Add device id to vfio driver.
+	api.DeviceLogger().WithFields(logrus.Fields{
+		"vendor-device-id": vendorDeviceID,
+		"vfio-new-id-path": vfioNewIDPath,
+	}).Info("Writing vendor-device-id to vfio new-id path")
+
+	if err := utils.WriteToFile(vfioNewIDPath, []byte(vendorDeviceID)); err != nil {
+		return err
+	}
+
+	// Bind to vfio-pci driver.
+	bindDriverPath := fmt.Sprintf(pciDriverBindPath, "vfio-pci")
+
+	api.DeviceLogger().WithFields(logrus.Fields{
+		"device-bdf":  bdf,
+		"driver-path": bindDriverPath,
+	}).Info("Binding device to vfio driver")
+
+	// Device may be already bound at this time because of earlier write to new_id, ignore error
+	utils.WriteToFile(bindDriverPath, []byte(bdf))
+
+	return nil
+}
+
+// bindDevicetoHost binds the device to the host driver driver after unbinding from vfio-pci.
+func bindDevicetoHost(bdf, hostDriver, vendorDeviceID string) error {
+	// Unbind from vfio-pci driver
+	unbindDriverPath := fmt.Sprintf(pciDriverUnbindPath, bdf)
+	api.DeviceLogger().WithFields(logrus.Fields{
+		"device-bdf":  bdf,
+		"driver-path": unbindDriverPath,
+	}).Info("Unbinding device from driver")
+
+	if err := utils.WriteToFile(unbindDriverPath, []byte(bdf)); err != nil {
+		return err
+	}
+
+	// To prevent new VFs from binding to VFIO-PCI, remove_id
+	if err := utils.WriteToFile(vfioRemoveIDPath, []byte(vendorDeviceID)); err != nil {
+		return err
+	}
+
+	// Bind back to host driver
+	bindDriverPath := fmt.Sprintf(pciDriverBindPath, hostDriver)
+	api.DeviceLogger().WithFields(logrus.Fields{
+		"device-bdf":  bdf,
+		"driver-path": bindDriverPath,
+	}).Info("Binding back device to host driver")
+
+	return utils.WriteToFile(bindDriverPath, []byte(bdf))
+}
+
 func bindNICToVFIO(endpoint *PhysicalEndpoint) error {
 	return bindDevicetoVFIO(endpoint.BDF, endpoint.Driver, endpoint.VendorDeviceID)
 }
 
 func bindNICToHost(endpoint *PhysicalEndpoint) error {
 	return bindDevicetoHost(endpoint.BDF, endpoint.Driver, endpoint.VendorDeviceID)
+}
+
+// Long term, this should be made more configurable.  For now matching path
+// provided by CNM VPP and OVS-DPDK plugins, available at github.com/clearcontainers/vpp and
+// github.com/clearcontainers/ovsdpdk.  The plugins create the socket on the host system
+// using this path.
+const hostSocketSearchPath = "/tmp/vhostuser_%s/vhu.sock"
+
+// findVhostUserNetSocketPath checks if an interface is a dummy placeholder
+// for a vhost-user socket, and if it is it returns the path to the socket
+func findVhostUserNetSocketPath(netInfo NetworkInfo) (string, error) {
+	if netInfo.Iface.Name == "lo" {
+		return "", nil
+	}
+
+	// check for socket file existence at known location.
+	for _, addr := range netInfo.Addrs {
+		socketPath := fmt.Sprintf(hostSocketSearchPath, addr.IPNet.IP)
+		if _, err := os.Stat(socketPath); err == nil {
+			return socketPath, nil
+		}
+	}
+
+	return "", nil
+}
+
+// vhostUserSocketPath returns the path of the socket discovered.  This discovery
+// will vary depending on the type of vhost-user socket.
+//  Today only VhostUserNetDevice is supported.
+func vhostUserSocketPath(info interface{}) (string, error) {
+
+	switch v := info.(type) {
+	case NetworkInfo:
+		return findVhostUserNetSocketPath(v)
+	default:
+		return "", nil
+	}
+
 }
 
 // network is the virtcontainers network interface.
