+++ virtcontainers/qemu.go
@@ -16,6 +16,10 @@ import (
 	govmmQemu "github.com/intel/govmm/qemu"
 	"github.com/kata-containers/runtime/virtcontainers/pkg/uuid"
 	"github.com/sirupsen/logrus"
+
+	"github.com/kata-containers/runtime/virtcontainers/device/api"
+	deviceDrivers "github.com/kata-containers/runtime/virtcontainers/device/drivers"
+	"github.com/kata-containers/runtime/virtcontainers/utils"
 )
 
 type qmpChannel struct {
@@ -593,7 +597,7 @@ func (q *qemu) removeDeviceFromBridge(ID string) error {
 	return err
 }
 
-func (q *qemu) hotplugBlockDevice(drive *Drive, op operation) error {
+func (q *qemu) hotplugBlockDevice(drive *deviceDrivers.Drive, op operation) error {
 	defer func(qemu *qemu) {
 		if q.qmpMonitorCh.qmp != nil {
 			q.qmpMonitorCh.qmp.Shutdown()
@@ -634,7 +638,7 @@ func (q *qemu) hotplugBlockDevice(drive *Drive, op operation) error {
 			bus := scsiControllerID + ".0"
 
 			// Get SCSI-id and LUN based on the order of attaching drives.
-			scsiID, lun, err := getSCSIIdLun(drive.Index)
+			scsiID, lun, err := utils.GetSCSIIdLun(drive.Index)
 			if err != nil {
 				return err
 			}
@@ -662,7 +666,7 @@ func (q *qemu) hotplugBlockDevice(drive *Drive, op operation) error {
 	return nil
 }
 
-func (q *qemu) hotplugVFIODevice(device VFIODevice, op operation) error {
+func (q *qemu) hotplugVFIODevice(device deviceDrivers.VFIODevice, op operation) error {
 	defer func(qemu *qemu) {
 		if q.qmpMonitorCh.qmp != nil {
 			q.qmpMonitorCh.qmp.Shutdown()
@@ -703,13 +707,15 @@ func (q *qemu) hotplugVFIODevice(device VFIODevice, op operation) error {
 func (q *qemu) hotplugDevice(devInfo interface{}, devType deviceType, op operation) error {
 	switch devType {
 	case blockDev:
-		drive := devInfo.(*Drive)
+		// TODO: find a way to remove dependency of deviceDrivers lib @weizhang555
+		drive := devInfo.(*deviceDrivers.Drive)
 		return q.hotplugBlockDevice(drive, op)
 	case cpuDev:
 		vcpus := devInfo.(uint32)
 		return q.hotplugCPUs(vcpus, op)
 	case vfioDev:
-		device := devInfo.(VFIODevice)
+		// TODO: find a way to remove dependency of deviceDrivers lib @weizhang555
+		device := devInfo.(deviceDrivers.VFIODevice)
 		return q.hotplugVFIODevice(device, op)
 	default:
 		return fmt.Errorf("cannot hotplug device: unsupported device type '%v'", devType)
@@ -840,6 +846,13 @@ func (q *qemu) resumeSandbox() error {
 
 // addDevice will add extra devices to Qemu command line.
 func (q *qemu) addDevice(devInfo interface{}, devType deviceType) error {
+	switch devType {
+	case vhostuserDev:
+		vhostDev := devInfo.(api.VhostUserDevice)
+		q.qemuConfig.Devices = q.arch.appendVhostUserDevice(q.qemuConfig.Devices, vhostDev)
+		return nil
+	}
+
 	switch v := devInfo.(type) {
 	case Volume:
 		q.qemuConfig.Devices = q.arch.append9PVolume(q.qemuConfig.Devices, v)
@@ -847,17 +860,10 @@ func (q *qemu) addDevice(devInfo interface{}, devType deviceType) error {
 		q.qemuConfig.Devices = q.arch.appendSocket(q.qemuConfig.Devices, v)
 	case Endpoint:
 		q.qemuConfig.Devices = q.arch.appendNetwork(q.qemuConfig.Devices, v)
-	case Drive:
+	case deviceDrivers.Drive:
 		q.qemuConfig.Devices = q.arch.appendBlockDevice(q.qemuConfig.Devices, v)
 
-	//vhostUserDevice is an interface, hence the pointer for Net, SCSI and Blk:
-	case VhostUserNetDevice:
-		q.qemuConfig.Devices = q.arch.appendVhostUserDevice(q.qemuConfig.Devices, &v)
-	case VhostUserSCSIDevice:
-		q.qemuConfig.Devices = q.arch.appendVhostUserDevice(q.qemuConfig.Devices, &v)
-	case VhostUserBlkDevice:
-		q.qemuConfig.Devices = q.arch.appendVhostUserDevice(q.qemuConfig.Devices, &v)
-	case VFIODevice:
+	case deviceDrivers.VFIODevice:
 		q.qemuConfig.Devices = q.arch.appendVFIODevice(q.qemuConfig.Devices, v)
 	default:
 		break
